+++ paste.bin.ingest.cmd/Constants.cs
@@ -8,8 +8,6 @@
         // data constants
         public const string PasteBinBaseUrl = "https://pastebin.com/archive";
         public const string PasteBinRawUrl = "https://pastebin.com/raw";
-        public const string DataSaveLocation = "E:\\pastebin";
-        public const string LogSaveLocation = "E:\\pastebin\\cli.logs";
 
         // logging constants:
         public const string InitLogger = "initialized logger";